Fixes autorotate double-delay on 1st and last slides
The earlier autorotate code appears to cause the first and last slide to wait for twice as long as the others before moving (i.e. if the data-autorotate value is set to '3000', the first and last slides will actually pause for 6000ms).

This code change also caches the $slider object and appears to fix an issue where the direction of movement for one carousel can affect the movement of other carousels on the page.